requirements.txt
streamlit
langchain
langchain-community
langchain-postgres
langchain-ollama
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
psycopg2-binary